Deploy Job resource to kubernetes
Add deploy Job in deploy pkg
