+++ pkg/create/createReplicationController.go
@@ -6,13 +6,12 @@ import (
 	"github.com/ghodss/yaml"
 	"github.com/weitenghuang/dirigent-cli/pkg/kubernetes/api"
 	"github.com/weitenghuang/dirigent-cli/pkg/resource"
-	"github.com/weitenghuang/dirigent-cli/pkg/utils"
 	"io/ioutil"
 	"k8s.io/kubernetes/pkg/api/unversioned"
 )
 
 func ReplicationController(appName string, appConfig *config.ServiceConfig) (string, error) {
-	if stop, err := utils.StopJobResourceWithError(appName); stop && err != nil {
+	if stop, err := resource.StopJobResourceWithError(appName); stop && err != nil {
 		return "", err
 	}
