+++ pkg/deploy/deployService.go
@@ -3,13 +3,14 @@ package deploy
 import (
 	"github.com/docker/libcompose/config"
 	"github.com/weitenghuang/dirigent-cli/pkg/create"
+	"github.com/weitenghuang/dirigent-cli/pkg/resource"
 	"github.com/weitenghuang/dirigent-cli/pkg/utils"
 	"os"
 )
 
 // func DeployService(appKey string, appValue map[string]interface{}) error {
 func Service(appName string, appConfig *config.ServiceConfig) error {
-	if stop, err := utils.StopJobResourceWithError(appName); stop && err != nil {
+	if stop, err := resource.StopJobResourceWithError(appName); stop && err != nil {
 		return err
 	}
